+++ tests/Profile/Shopware/Gateway/Local/LanguageReaderTest.php
@@ -7,6 +7,7 @@
 
 namespace SwagMigrationAssistant\Test\Profile\Shopware\Gateway\Local;
 
+use Doctrine\DBAL\Connection;
 use PHPUnit\Framework\TestCase;
 use Shopware\Core\Framework\Log\Package;
 use SwagMigrationAssistant\Migration\MigrationContext;
@@ -25,11 +26,18 @@ class LanguageReaderTest extends TestCase
 
     private MigrationContext $migrationContext;
 
+    private Connection $dbConnection;
+
+    private int $customerLocaleId = 9999;
+
+    private bool $customerLocaleInserted = false;
+
     protected function setUp(): void
     {
         $this->connectionSetup();
 
-        $this->languageReader = new LanguageReader(new ConnectionFactory());
+        $connectionFactory = new ConnectionFactory();
+        $this->languageReader = new LanguageReader($connectionFactory);
 
         $this->migrationContext = new MigrationContext(
             $this->connection,
@@ -42,6 +50,25 @@ class LanguageReaderTest extends TestCase
         );
 
         $this->migrationContext->setGateway(new DummyLocalGateway());
+
+        $this->dbConnection = $connectionFactory->createDatabaseConnection($this->migrationContext);
+        $this->dbConnection->executeStatement('UPDATE s_user SET language = :language WHERE id = :id', [
+            'language' => 1,
+            'id' => 1,
+        ]);
+    }
+
+    protected function tearDown(): void
+    {
+        // reset customer to the fixture default locale
+        $this->dbConnection->executeStatement('UPDATE s_user SET language = :language WHERE id = :id', [
+            'language' => 1,
+            'id' => 1,
+        ]);
+
+        if ($this->customerLocaleInserted) {
+            $this->dbConnection->executeStatement('DELETE FROM s_core_locales WHERE id = :id', ['id' => $this->customerLocaleId]);
+        }
     }
 
     public function testRead(): void
@@ -62,5 +89,39 @@ class LanguageReaderTest extends TestCase
         static::assertSame('de-DE', $data[1]['_locale']);
         static::assertSame('de_DE', $data[1]['translations'][0]['locale']);
         static::assertSame('en_GB', $data[1]['translations'][1]['locale']);
+    }
+
+    public function testReadIncludesCustomerLocalesOutsideShopLocales(): void
+    {
+        $existingLocaleId = $this->dbConnection->fetchOne(
+            'SELECT id FROM s_core_locales WHERE locale = :locale',
+            ['locale' => 'ar_EG']
+        );
+
+        if ($existingLocaleId === false) {
+            $this->dbConnection->executeStatement(
+                'INSERT INTO s_core_locales (id, locale, language, territory) VALUES (:id, :locale, :language, :territory)',
+                [
+                    'id' => $this->customerLocaleId,
+                    'locale' => 'ar_EG',
+                    'language' => 'Arabic',
+                    'territory' => 'Egypt',
+                ]
+            );
+
+            $this->customerLocaleInserted = true;
+        } else {
+            $this->customerLocaleId = (int) $existingLocaleId;
+        }
+
+        $this->dbConnection->executeStatement('UPDATE s_user SET language = :language WHERE id = :id', [
+            'language' => $this->customerLocaleId,
+            'id' => 1,
+        ]);
+
+        $data = $this->languageReader->read($this->migrationContext);
+
+        static::assertCount(3, $data);
+        static::assertContains('ar-EG', \array_column($data, 'locale'));
     }
 }