+++ tests/Profile/Shopware55/Converter/CustomerConverterTest.php
@@ -372,4 +372,74 @@ class CustomerConverterTest extends TestCase
         static::assertSame('Shopware AG', $converted['company']);
         static::assertSame(CustomerEntity::ACCOUNT_TYPE_BUSINESS, $converted['accountType']);
     }
+
+    public function testConvertKeepsDefaultAddressesWhenAddressSalutationIsUnknown(): void
+    {
+        $customerData = require __DIR__ . '/../../../_fixtures/customer_data.php';
+        $customerData = $customerData[0];
+        $customerData['addresses'][0]['salutation'] = 'unknown-salutation';
+
+        $context = Context::createDefaultContext();
+        $convertResult = $this->customerConverter->convert(
+            $customerData,
+            $context,
+            $this->migrationContext
+        );
+
+        $converted = $convertResult->getConverted();
+        static::assertNotNull($converted);
+        static::assertCount(3, $converted['addresses']);
+        static::assertArrayNotHasKey('salutationId', $converted['addresses'][0]);
+        static::assertSame($converted['addresses'][0]['id'], $converted['defaultBillingAddressId']);
+        static::assertSame($converted['addresses'][1]['id'], $converted['defaultShippingAddressId']);
+
+        $logs = $this->loggingService->getLoggingArray();
+        static::assertCount(1, $logs);
+        static::assertSame(ConvertEntityUnknownLog::getCode(), $logs[0]['code']);
+    }
+
+    public function testConvertUsesLanguageMappingWhenLanguageDoesNotExistYet(): void
+    {
+        $customerData = require __DIR__ . '/../../../_fixtures/customer_data.php';
+        $customerData = $customerData[0];
+        $customerData['customerlanguage']['locale'] = 'ar-EG';
+
+        $languageId = Uuid::randomHex();
+        $context = Context::createDefaultContext();
+        $this->mappingService->getOrCreateMapping(
+            $this->connectionId,
+            DefaultEntities::LANGUAGE,
+            'ar-EG',
+            $context,
+            null,
+            [],
+            $languageId
+        );
+
+        $languageLookup = $this->createMock(LanguageLookup::class);
+        $languageLookup->expects($this->never())->method('get');
+
+        $validator = static::getContainer()->get('validator');
+        $salesChannelRepo = static::getContainer()->get('sales_channel.repository');
+
+        $customerConverter = new Shopware55CustomerConverter(
+            $this->mappingService,
+            $this->loggingService,
+            $validator,
+            $salesChannelRepo,
+            static::getContainer()->get(CountryLookup::class),
+            $languageLookup,
+            static::getContainer()->get(CountryStateLookup::class),
+        );
+
+        $convertResult = $customerConverter->convert(
+            $customerData,
+            $context,
+            $this->migrationContext
+        );
+
+        $converted = $convertResult->getConverted();
+        static::assertNotNull($converted);
+        static::assertSame($languageId, $converted['languageId']);
+    }
 }